move all classes from 'universal' module to 'android' module
