+++ android/build.gradle
@@ -68,7 +68,6 @@ task sourceJar(type: Jar) {
 }
 
 dependencies {
-    implementation project(':universal')
     implementation 'net.freehaven.tor.control:jtorctl:0.2'
     implementation 'org.slf4j:slf4j-api:1.7.30'
     implementation 'org.slf4j:slf4j-android:1.7.30'
